// src/types/blog.ts
export interface Blog {
  id: string
  title: string
  category: string[]
  description: string
  date: string
  coverImage: string
  content: string
}

export interface CreateBlogInput {
  title: string
  category: string[]
  description: string
  date: string
  coverImage: string
  content: string
}